Test new DynamoDb testing package.
