+++ AssetInformationApi.Tests/DynamoDbCollection.cs
@@ -1,0 +1,12 @@
+using Xunit;
+
+namespace AssetInformationApi.Tests
+{
+    [CollectionDefinition("DynamoDb collection", DisableParallelization = true)]
+    public class DynamoDbCollection : ICollectionFixture<MockWebApplicationFactory<Startup>>
+    {
+        // This class has no code, and is never created. Its purpose is simply
+        // to be the place to apply [CollectionDefinition] and all the
+        // ICollectionFixture<> interfaces.
+    }
+}
