--- AssetInformationApi.Tests/DynamoDbMockWebApplicationFactory.cs
@@ -1,90 +1,0 @@
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.DataModel;
-using Amazon.DynamoDBv2.Model;
-using Hackney.Core.DynamoDb;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace AssetInformationApi.Tests
-{
-    public class DynamoDbMockWebApplicationFactory<TStartup>
-        : WebApplicationFactory<TStartup> where TStartup : class
-    {
-        private readonly List<TableDef> _tables;
-
-        public IAmazonDynamoDB DynamoDb { get; private set; }
-        public IDynamoDBContext DynamoDbContext { get; private set; }
-
-        public DynamoDbMockWebApplicationFactory(List<TableDef> tables)
-        {
-            _tables = tables;
-        }
-
-        protected override void ConfigureWebHost(IWebHostBuilder builder)
-        {
-            builder.ConfigureAppConfiguration(b => b.AddEnvironmentVariables())
-                .UseStartup<Startup>();
-            builder.ConfigureServices(services =>
-            {
-                services.ConfigureDynamoDB();
-
-                var serviceProvider = services.BuildServiceProvider();
-                DynamoDb = serviceProvider.GetRequiredService<IAmazonDynamoDB>();
-                DynamoDbContext = serviceProvider.GetRequiredService<IDynamoDBContext>();
-
-                EnsureTablesExist(DynamoDb, _tables);
-            });
-        }
-
-        private static void EnsureTablesExist(IAmazonDynamoDB dynamoDb, List<TableDef> tables)
-        {
-            foreach (var table in tables)
-            {
-                try
-                {
-                    var keySchema = new List<KeySchemaElement> { new KeySchemaElement(table.KeyName, KeyType.HASH) };
-                    var attributes = new List<AttributeDefinition> { new AttributeDefinition(table.KeyName, table.KeyType) };
-                    if (!string.IsNullOrEmpty(table.RangeKeyName))
-                    {
-                        keySchema.Add(new KeySchemaElement(table.RangeKeyName, KeyType.RANGE));
-                        attributes.Add(new AttributeDefinition(table.RangeKeyName, table.RangeKeyType));
-                    }
-
-                    foreach (var localIndex in table.LocalSecondaryIndexes)
-                    {
-                        var indexRangeKey = localIndex.KeySchema.FirstOrDefault(y => y.KeyType == KeyType.RANGE);
-                        if ((null != indexRangeKey) && (!attributes.Any(x => x.AttributeName == indexRangeKey.AttributeName)))
-                            attributes.Add(new AttributeDefinition(indexRangeKey.AttributeName, ScalarAttributeType.S)); // Assume a string for now.
-                    }
-
-                    foreach (var globalIndex in table.GlobalSecondaryIndexes)
-                    {
-                        foreach (var key in globalIndex.KeySchema)
-                        {
-                            if (!attributes.Any(x => x.AttributeName == key.AttributeName))
-                                attributes.Add(new AttributeDefinition(key.AttributeName, ScalarAttributeType.S)); // Assume a string for now.
-                        }
-                    }
-
-                    var request = new CreateTableRequest(table.Name,
-                        keySchema,
-                        attributes,
-                        new ProvisionedThroughput(3, 3))
-                    {
-                        LocalSecondaryIndexes = table.LocalSecondaryIndexes,
-                        GlobalSecondaryIndexes = table.GlobalSecondaryIndexes
-                    };
-                    _ = dynamoDb.CreateTableAsync(request).GetAwaiter().GetResult();
-                }
-                catch (ResourceInUseException)
-                {
-                    // It already exists :-)
-                }
-            }
-        }
-    }
-}
