+++ AssetInformationApi.Tests/LogCallAspectFixtureCollection.cs
@@ -1,0 +1,13 @@
+using Hackney.Core.Testing.Shared;
+using Xunit;
+
+namespace AssetInformationApi.Tests
+{
+    [CollectionDefinition("LogCall collection")]
+    public class LogCallAspectFixtureCollection : ICollectionFixture<LogCallAspectFixture>
+    {
+        // This class has no code, and is never created. Its purpose is simply
+        // to be the place to apply [CollectionDefinition] and all the
+        // ICollectionFixture<> interfaces.
+    }
+}
