+++ AssetInformationApi.Tests/V1/E2ETests/Fixtures/AssetsFixture.cs
@@ -1,5 +1,5 @@
-using Amazon.DynamoDBv2.DataModel;
 using AutoFixture;
+using Hackney.Core.Testing.DynamoDb;
 using Hackney.Shared.Asset.Infrastructure;
 using System;
 
@@ -8,15 +8,15 @@ namespace AssetInformationApi.Tests.V1.E2ETests.Fixtures
     public class AssetsFixture : IDisposable
     {
         private readonly Fixture _fixture = new Fixture();
-        private readonly IDynamoDBContext _dbContext;
+        private readonly IDynamoDbFixture _dbFixture;
 
         public AssetDb Asset { get; private set; }
         public Guid AssetId { get; private set; }
         public string InvalidAssetId { get; private set; }
 
-        public AssetsFixture(IDynamoDBContext dbContext)
+        public AssetsFixture(IDynamoDbFixture dbFixture)
         {
-            _dbContext = dbContext;
+            _dbFixture = dbFixture;
         }
 
         public void Dispose()
@@ -31,7 +31,7 @@ namespace AssetInformationApi.Tests.V1.E2ETests.Fixtures
             if (disposing && !_disposed)
             {
                 if (Asset != null)
-                    _dbContext.DeleteAsync(Asset).GetAwaiter().GetResult();
+                    _dbFixture.DynamoDbContext.DeleteAsync(Asset).GetAwaiter().GetResult();
 
                 _disposed = true;
             }
@@ -45,7 +45,7 @@ namespace AssetInformationApi.Tests.V1.E2ETests.Fixtures
 
             AssetId = Asset.Id;
 
-            _dbContext.SaveAsync(Asset).GetAwaiter().GetResult();
+            _dbFixture.DynamoDbContext.SaveAsync(Asset).GetAwaiter().GetResult();
         }
 
         public void GivenAnAssetThatDoesntExist()
